Renomme les fichiers des dépôts de données homologations et services
